--- config.json
@@ -1,31 +1,0 @@
-{
-    "port": 3002,
-    "accepts": ["application/vnd.oasis.taxii+json; version=2.0", "application/vnd.oasis.taxii+json", "application/vnd.oasis.stix+json; version=2.0", "application/vnd.oasis.stix+json"],
-    "bind_address": "0.0.0.0",
-    "connection_string": "mongodb://repository:27017/",
-    "response_type": {"taxii": "application/vnd.oasis.taxii+json; version=2.0", "stix": "application/vnd.oasis.stix+json; version=2.0"},
-    "bundle_spec_version": "2.0",
-    "discovery": {
-            "title": "CTI TAXII server",
-            "description": "This TAXII server contains a listing of ATT&CK domain collections expressed as STIX, including PRE-ATT&CK, ATT&CK for Enterprise, and ATT&CK Mobile.",
-            "contact": "attack@mitre.org",
-            "default": "stix",
-            "api_roots": [
-                "stix"
-            ]
-    },
-    "autogenerate_roots": {
-        "enabled": 1,
-        "versions": ["taxii-2.0"],
-        "max_content_length": "10485760"
-    },
-    "roots": [
-        {
-            "title": "Default API root",
-            "description": "This API root contains...",
-            "versions": ["taxii-2.0"],
-            "max_content_length": "10485760",
-            "path": "stix"
-        }
-    ]
-}
