--- models/object.js
@@ -1,6 +1,0 @@
-const mongoose = require('../init');
-const Schema = mongoose.Schema;
-
-const ObjectSchema = new Schema({ any: {} });
-
-module.exports = ObjectSchema;
