+++ package.json
@@ -16,18 +16,19 @@
   "license": "ISC",
   "dependencies": {
     "bluebird": "^3.5.1",
-    "chai": "^4.1.2",
-    "chai-http": "^3.0.0",
-    "eslint": "^4.13.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
     "express": "^4.16.2",
     "express-paginate": "^0.3.0",
-    "mocha": "^4.0.1",
     "mongoose": "^4.12.5",
     "uuid": "^3.2.1",
     "uuid-validate": "0.0.2"
+  },
+  "devDependencies": {
+    "@types/express": "^4.11.1",
+    "@types/mongoose": "^5.0.10",
+    "@types/uuid": "^3.4.3",
+    "@types/uuid-validate": "0.0.0",
+    "concurrently": "^3.5.1",
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.3"
   }
 }
