+++ src/errors/app-error.ts
@@ -1,7 +1,7 @@
-module.exports = class AppError extends Error {
-  constructor(message) {
+export default class AppError extends Error {
+  constructor(message: any) {
     super(message);
     this.name = this.constructor.name;
     Error.captureStackTrace(this, this.constructor);
   }
-};
+}
