+++ src/errors/invalid-uuid.ts
@@ -1,0 +1,9 @@
+import AppError from './app-error';
+
+export default class extends AppError {
+  public fields: any;
+  constructor(fields: any) {
+    super('Invalid UUID present in collections file');
+    this.fields = fields || {};
+  }
+}
