+++ src/models/object.ts
@@ -1,0 +1,7 @@
+import init from '../server/init';
+
+const Schema = init.Schema;
+
+const ObjectSchema = new Schema({ any: {} });
+
+export default ObjectSchema;
